Add support for diff syntax highlighting
In Nova 11 (currently in beta), the built-in diff syntax highlighting
will come via Tree-sitter, meaning I can use it here as a injected
sub-language. This will allow the diff portion of a Git commit message
to be highlighted too.

This should be compatible with Nova 10 --- if I remember correctly, the
diff portion of a commit message will simply be highlighted as a
comment. (In Nova 10, the diff syntax is provided as a regex grammar,
which isn't compatible with a Tree-sitter grammar. It fails silently.)

https://devforum.nova.app/t/nova-11-beta-1-released/2033
https://devforum.nova.app/t/could-novas-built-in-diff-syntax-be-provided-as-a-tree-sitter-grammar/2001